requirements.txt
pandas>=2.0.0
numpy>=1.24.0
requests>=2.31.0
openpyxl>=3.1.0
pyarrow>=14.0.0
office365-rest-python-client>=2.5.0
